refactor(assistant-stream): share incremental JSON parser
