+++ .changeset/incremental-langgraph-tool-args.md
@@ -1,5 +1,6 @@
 ---
 "@assistant-ui/react-langgraph": patch
+"assistant-stream": patch
 ---
 
-fix: parse streamed LangGraph tool arguments incrementally instead of reparsing the complete prefix for every delta
+fix: add shared incremental partial-JSON parsing and use it for streamed LangGraph tool arguments
